test(#365) : 어드민 컨트롤러 분리
